게시판 글쓰기처리 및 이미지 DB연결 필요

+++ src/main/webapp/WEB-INF/views/board/ann_selectone.jsp
@@ -48,7 +48,7 @@
 
             <div class="bt_wrap">
                 <a href="announce_update.do?num=${param.num}">수정</a>
-                <a href="announce_delete.do?num=${param.num}">삭제</a>
+                <a href="announce_delete.do?num=${param.num}" onclick="return confirm('삭제하시겠습니까?')">삭제</a>
             </div>
     
 </body>
